requirements.txt
mail-parser
imapclient
pyyaml